Fix build break on amd64 platforms - Pointer truncation warning removed - Use uintptr_t instead of int
[tfs-changeset: 1647853]

+++ src/Native/Runtime/TypeManager.cpp
@@ -171,7 +171,7 @@ HANDLE TypeManager::GetOsModuleHandle()
 bool TypeManagerHandle::IsTypeManager()
 {
 #if !CORERT
-    if (((int)_value & 1) == 0)
+    if (((uintptr_t)_value & 1) == 0)
         return false;
 #endif
